Working on addressing the f64s in the data
Currently, only the X coeffs can be read. It's an offset issue which
shouldn't be too hard to fix, but I'm too tired.

Also, new file is only 2.3% smaller than SPK. This is not surprising:
most of the savings were in removing the extra 8 octets of the header of
the coefficients, and this refactoring adds one back for Chebyshev
evenly spaced splines.