Converting LUT to DataArray fails
I don't know how to allow a Vec to move into a large lifetime...
There is a big issue when building a DataArray: we don't know how many items exist, so we need an allocator.

If the solution is to allow for allocs when building, that's OK, but
then we can't easily read data from SPICE without allocating before reading.

The only good thing here is that we use very little stack space for passing the data around

Signed-off-by: Christopher Rabotin <christopher.rabotin@gmail.com>