Switched the underlying of DAF to `bytes::Bytes`
This removes the need for a lifetime in DAF.
It also allows for a DAF to be loaded both from a borrowed byte array and an owned byte array, all while being tiny tiny tiny in stack.

The biggest drawback so far is that the data summaries must be parsed on the fly every time. I'm sure there's a better design but I haven't thought of it yet.

I still need to fix the mmap call.

Signed-off-by: Christopher Rabotin <christopher.rabotin@gmail.com>