Can now parse a KPL/PCK
Now I need to build a structure to store it because the parsing only
works with std and alloc. I think that ASN1 would be perfect for this.
Essentially, any user with std will be able to read a KPL PCK (.tpc)
file without issue. That would then create an ASN1 stored in memory
(which could be serialized and loaded onboard).