Type limitation in Dhall and Python
VectorExpr should allow building any recursive vector. This would
extend to building a projection VectorExpr, e.g.
VectorExpr::Project {vector: Box::new(VectorExpr), plane: PlaneXY, of: (StateSpec, FrameSpec::LocalVNC)}

This cannot be exported to Python because of the Box<Self>. So it would need to be serialized...
but that also isn't yet supported in Dhall.

One option is to disable that variant when building the Python bindings and creating a whole new
initializer for it. But that isn't very clean, and I also would need two different enums
with the same name. Bref, it wouldn't work.