feat: all underlying DAF data is now a BytesMut instead of a Bytes
Also removed the MMap in file2heap: mmap errors are interrupt errors that can't be caught. ANISE was still loading everything on the heap, so there way zero advantage of loading the file via mmap anyway.
The CRC32 is also not computed by default when loading the DAF files: if users need the CRC for scrubbing, they must request for it to be set. That should shave off a tiny bit of time in loading large files.